# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hex Microservices API
  description: |
    Comprehensive API documentation for the Hex Microservices architecture.
    
    ## Architecture Overview
    - **API Gateway** (port 3000): Authentication and routing
    - **Hex Data Service** (port 3001): TCP data logging and management
    - **Notification Service** (port 3002): Push notifications and device management
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain tokens via `/api/auth/login`.
    
    ## Base URLs
    - API Gateway: `http://localhost:3000`
    - Hex Data Service: `http://localhost:3001` 
    - Notification Service: `http://localhost:3002`
  version: 1.0.0
  contact:
    name: Hex Microservices API
    email: developer@hexmicroservices.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: API Gateway (main entry point)
  - url: http://localhost:3001
    description: Hex Data Service
  - url: http://localhost:3002
    description: Notification Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message (present when success is false)
        message:
          type: string
          description: Additional information about the response
      required:
        - success
...
    DeviceToken:
      type: object
      properties:
        id:
          type: string
          description: Unique device identifier
        token:
          type: string
          description: Push notification token
        platform:
          type: string
          enum: [ios, android, web]
          description: Device platform
        userId:
          type: string
          description: Associated user ID (optional)
        createdAt:
          type: string
          format: date-time
          description: When the device was registered
        isActive:
          type: boolean
          description: Whether the device is active
      required:
        - id
        - token
        - platform
        - createdAt
        - isActive
...
paths:
...
  /api/devices/register:
    post:
      tags:
        - Devices
      summary: Register a new device
      description: Register a device token for push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Push notification token
                platform:
                  type: string
                  enum: [ios, android, web]
                  description: Device platform
                userId:
                  type: string
                  description: Associated user ID (optional)
              required:
                - token
                - platform
            example:
              token: "fcm-token-abc123"
              platform: "android"
              userId: "user-123"
      responses:
        '201':
          description: Device registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DeviceToken'
        '400':
          description: Invalid platform or missing required fields
  /api/devices/{deviceId}:
    get:
      tags:
        - Devices
      summary: Get device by ID
      description: Retrieve a specific device by its ID
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
          description: Device ID
      responses:
        '200':
          description: Device found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DeviceToken'
        '404':
          description: Device not found

    put:
      tags:
        - Devices
      summary: Update device
      description: Update device information (cannot update ID or token)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platform:
                  type: string
                  enum: [ios, android, web]
                userId:
                  type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DeviceToken'
        '404':
          description: Device not found

    delete:
      tags:
        - Devices
      summary: Remove device
      description: Permanently remove a device from the system
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Device not found
  /api/devices/{deviceId}/deactivate:
    delete:
      tags:
        - Devices
      summary: Deactivate device
      description: Deactivate a device (mark as inactive without removing)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Device not found
  /api/devices/user/{userId}:
    get:
      tags:
        - Devices
      summary: Get devices for a user
      description: Get all devices registered for a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: User devices retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          devices:
                            type: array
                            items:
                              $ref: '#/components/schemas/DeviceToken'
                          count:
                            type: integer
  /api/devices/platform/{platform}:
    get:
      tags:
        - Devices
      summary: Get devices by platform
      description: Get all devices for a specific platform
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [ios, android, web]
      responses:
        '200':
          description: Platform devices retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          devices:
                            type: array
                            items:
                              $ref: '#/components/schemas/DeviceToken'
                          count:
                            type: integer
                          platform:
                            type: string
...
  /api/devices/token/{token}:
    delete:
      tags:
        - Devices
      summary: Remove device by token
      description: Remove a device using its push notification token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Device push notification token
      responses:
        '200':
          description: Device removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Device not found
...
  /api/devices/admin/cleanup:
    post:
      tags:
        - Devices
      summary: Clean up inactive devices
      description: Remove inactive devices older than specified days
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  default: 30
                  description: Remove devices inactive for this many days
      responses:
        '200':
          description: Cleanup completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          removedCount:
                            type: integer